+++ packages/atomic/cypress/e2e/facets/numeric-facet/numeric-facet.cypress.ts
@@ -731,7 +731,7 @@ describe('Numeric Facet V1 Test Suites', () => {
       new TestFixture()
         .with(
           addNumericFacet({
-            field: 'dafsfs',
+            field: 'thisfielddoesnotexist',
             label: numericFacetLabel,
           })
         )