+++ packages/atomic/cypress/e2e/refine-toggle-actions.ts
@@ -21,7 +21,15 @@ import {createTimeframeElements} from './facets/timeframe-facet/timeframe-facet-
 import {timeframeFacetComponent} from './facets/timeframe-facet/timeframe-facet-selectors';
 import {refineToggleComponent} from './refine-toggle-selectors';
 
-export const addRefineToggle =
+export const addRefineToggleWithoutFacets =
+  (props: TagProps = {}) =>
+  (env: TestFixture) => {
+    const refineToggle = generateComponentHTML(refineToggleComponent, props);
+
+    env.withElement(refineToggle);
+  };
+
+export const addRefineToggleWithStaticFacets =
   (props: TagProps = {}) =>
   (env: TestFixture) => {
     const manager = generateComponentHTML(facetManagerComponent);
@@ -55,14 +63,6 @@ export const addRefineToggle =
     env.withElement(manager).withElement(refineToggle);
   };
 
-export const addRefineToggleWithoutFacets =
-  (props: TagProps = {}) =>
-  (env: TestFixture) => {
-    const refineToggle = generateComponentHTML(refineToggleComponent, props);
-
-    env.withElement(refineToggle);
-  };
-
 export const addRefineToggleWithAutomaticFacets =
   (props: TagProps = {}) =>
   (env: TestFixture) => {
@@ -73,6 +73,29 @@ export const addRefineToggleWithAutomaticFacets =
     const refineToggle = generateComponentHTML(refineToggleComponent, props);
 
     env.withElement(automaticFacetGenerator).withElement(refineToggle);
+  };
+
+export const addFacetManagerWithBothTypesOfFacets =
+  (props: TagProps = {}) =>
+  (env: TestFixture) => {
+    const manager = generateComponentHTML(facetManagerComponent);
+    manager.append(generateComponentHTML(facetComponent, {field: facetField}));
+    manager.append(
+      generateComponentHTML(numericFacetComponent, {field: numericFacetField})
+    );
+    manager.append(
+      generateComponentHTML(categoryFacetComponent, {field: hierarchicalField})
+    );
+    const automaticFacetGenerator = generateComponentHTML(
+      automaticFacetGeneratorComponent,
+      {'desired-count': '3'}
+    );
+    const refineToggle = generateComponentHTML(refineToggleComponent, props);
+
+    env
+      .withElement(manager)
+      .withElement(automaticFacetGenerator)
+      .withElement(refineToggle);
   };
 
 export const addRefineToggleRangeVariations =