+++ packages/headless/src/api/analytics/insight-analytics.ts
@@ -22,6 +22,10 @@ import {
 } from '../../state/state-sections';
 import {PreprocessRequest} from '../preprocess-request';
 import {BaseAnalyticsProvider} from './base-analytics';
+import {
+  wrapAnalyticsClientSendEventHook,
+  wrapPreprocessRequest,
+} from './coveo-analytics-utils';
 
 export type StateNeededByInsightAnalyticsProvider = ConfigurationSection &
   Partial<InsightAppState> &
@@ -116,9 +120,9 @@ export const configureInsightAnalytics = ({
       token,
       endpoint: apiBaseUrl,
       runtimeEnvironment,
-      preprocessRequest,
+      preprocessRequest: wrapPreprocessRequest(logger, preprocessRequest),
       beforeSendHooks: [
-        analyticsClientMiddleware,
+        wrapAnalyticsClientSendEventHook(logger, analyticsClientMiddleware),
         (type, payload) => {
           logger.info(
             {