+++ packages/headless/src/app/search-engine/jwt-reducer.ts
@@ -1,6 +1,6 @@
 import {isNullOrUndefined} from '@coveo/bueno';
 import {createReducer, Reducer} from '@reduxjs/toolkit';
-import {atob as atobshim} from 'abab';
+import {atob as atobShim} from 'abab';
 import P, {Logger} from 'pino';
 import {SearchAppState} from '../..';
 import {
@@ -70,7 +70,7 @@ const shouldReconcileValues = (
 
 const decodeJSONWebToken = (token: string): CoveoJSONWebToken | false => {
   try {
-    const atobImplementation = typeof atob !== 'undefined' ? atob : atobshim;
+    const atobImplementation = typeof atob !== 'undefined' ? atob : atobShim;
     const base64Url = token.split('.')[1];
     const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
     const base64decoded = atobImplementation(base64);